Improve progress indicators [Fixes #21]
Merged Skeleton components and progress bar indicators functionality.
For sync services, reports will work the same as before by just showing
the pulsing Skeleton components (which can be configured per-card to
match contents). For async services, the same Skeleton will be displayed
but a progress bar will also appear at the bottom of the Card.

These two components work well together now. The progress bar only runs
once, so if the report is taking longer than expected the pulsing
Skeleton contents signal that the report is still in progress, and is
not as frustrating (or inaccurate) as repeating the animation.

@danyocum I'm assuming the estimate set on the task is static when it is
created? I still consider this bit incomplete since for multi-minute
long-running tasks a user might close the window and come back to it. In
that case, the progress bar would just restart at the same position,
I think? If someone comes back to a report that takes 5 minutes, and
it's already been running for 4:40 the progress bar ought to display
that. Might not be super relevant now but for projects like MPATT and
future containerized jobs it will be more useful.